game/shared: clean up AI path position lines; add Minz/Maxz for size_t (#893)
* game: introduce local variables to clean up long AI path position lines

Break up the 200-char G_Ai_Node_GetPosition(VectorValue(...)) chains in
g_ai_goal.c and g_ai_main.c into readable intermediates. Also fixes the
implicit size_t -> int32_t/uint32_t precision warnings in those files and
in g_ai_node.c (three uncasted count assignments).


* shared: add Minz/Maxz for size_t; replace Mini casts at size_t call sites

Mini/Maxi operate on int32_t, requiring callers with size_t Vector counts
to add potentially lossy casts just to silence warnings. Minz/Maxz give
those call sites a purpose-built alternative with the correct type.

Update the three call sites in g_ai_goal.c, g_ai_main.c, and g_entity.c
that were casting ->count down to int32_t purely to satisfy Mini.


* client: use Minz in cl_server.c for size_t string length clamp


---------

+++ src/game/default/g_ai_node.c
@@ -163,7 +163,7 @@ static bool G_Ai_Node_EnsureSpatialIndex(void) {
 
   G_Ai_Node_InvalidateSpatialIndex();
 
-  const uint32_t n = g_ai_nodes->count;
+  const uint32_t n = (uint32_t) g_ai_nodes->count;
   g_ai_nodes_kdtree_positions = malloc(sizeof(vec3_t) * n);
   if (!g_ai_nodes_kdtree_positions) {
     return false;
@@ -1149,7 +1149,7 @@ void G_Ai_InitNodes(void) {
     const ai_node_t *node = AI_NODE(g_ai_nodes, i);
 
     if (node->links) {
-      g_ai_player_roam.file_links += node->links->count;
+      g_ai_player_roam.file_links += (uint32_t) node->links->count;
     }
   }
 }
@@ -1194,7 +1194,7 @@ void G_Ai_NodesReady(void) {
     return;
   }
 
-  const uint32_t added_nodes = g_ai_nodes->count - g_ai_player_roam.file_nodes;
+  const uint32_t added_nodes = (uint32_t) g_ai_nodes->count - g_ai_player_roam.file_nodes;
   uint32_t added_links = 0;
 
   for (uint32_t i = 0; i < g_ai_nodes->count; i++) {
